bigquery_query_to_table_src/requirements.txt
functions-framework==3.3.0
requests==2.28.2
google-cloud-bigquery==3.9.0
